refactor(core,mls,protocol): one sealed layer, shared with the leaf
A phone and a leaf node run different MLS implementations (ADR 0021) and
are not allowed to disagree about anything outside them. Four things sat
in that gap, all of them in crates that need `std`: the `EncryptedMessage`
envelope and its compact codec, `derive_address`, the domain-separated
canonical signing payload, and the sender-ratchet bounds.

They move to `offline-protocol-sealed`, a new dual std/no_std crate
between core and mls. Everything else re-exports or delegates, so no call
site, wire byte, error string, FFI signature or UDL entry changes. The
embedded footprint confirms it: the protocol layer is still 97,152 bytes.

A new crate rather than a core feature, because core links zero crypto and
carries no MLS vocabulary, and putting a hash there would reach a
relay-only image that never derives an address. See ADR 0022.

`tools/mls-interop` carried its own copies of the derivation rule and the
32/1000 ratchet constants, with a manifest comment saying nothing pinned
them. It now imports both, and a guard test reads the harness source and
fails if the copies come back: the harness is its own cargo workspace, so
that test is the only thing in `cargo test --workspace` that can see it.

Notes on the error types. `SealedError` repeats the `Display` strings of
the `MlsError` variants it replaced and is deliberately not
`#[non_exhaustive]`, so the mapping back is exhaustive; a wildcard arm is
how a new variant silently renders as another error's text (ADR 0013).
`FieldTooLarge` is its own variant because the engine renders it bare
while the MLS crate wraps it, and both strings had to stay identical.
